Сделан проект

+++ router.js
@@ -1,35 +1,28 @@
 import {MainPage} from "./src/components/main-page";
 import {FileUtils} from "./src/utils/file-utils";
-import {ExpenseList} from "./src/components/category expense/expense-list";
-import {ExpenseCreate} from "./src/components/category expense/expense-create";
-import {ExpenseEdit} from "./src/components/category expense/expense-edit";
-import {IncomeList} from "./src/components/category income/income-list";
-import {IncomeCreate} from "./src/components/category income/income-create";
-import {IncomeEdit} from "./src/components/category income/income-edit";
 import {OperationList} from "./src/components/operations/operation-list";
 import {OperationCreate} from "./src/components/operations/operation-create";
 import {OperationEdit} from "./src/components/operations/operation-edit";
-import {OperationDelete} from "./src/components/operations/operation-delete";
 import {SignUp} from "./src/components/auth/sign-up";
 import {Logout} from "./src/components/auth/logout";
 import {Login} from "./src/components/auth/login";
 import {AuthUtils} from "./src/utils/auth-utils";
+import {HttpUtils} from "./src/utils/http-utils";
+import {CategoryCreate} from "./src/components/category/category-create";
+import {CategoryEdit} from "./src/components/category/category-edit";
+import {CategoryList} from "./src/components/category/category-list";
 
 export class Router {
     constructor() {
         this.titlePageElement = document.getElementById('title');
         this.contentPageElement = document.getElementById('content');
-        this.userName = null;
-        this.userBalance = null;
 
         this.routes = [
             {
                 route: '/',
                 title: 'Главная',
                 filePathTemplate: '/templates/pages/main-page.html',
-                load: () => {
-                    new MainPage();
-                },
+                load: () => new MainPage(this.openNewRoute.bind(this)),
                 useLayout: '/templates/layout.html',
                 scripts: [
                     'chart.umd.js'
@@ -39,88 +32,64 @@ export class Router {
                 route: '/categories/income',
                 title: 'Категория доходов',
                 filePathTemplate: '/templates/pages/category income/list.html',
-                load: () => {
-                    new IncomeList(this.openNewRoute.bind(this));
-                },
+                load: () => new CategoryList(this.openNewRoute.bind(this), 'income'),
                 useLayout: '/templates/layout.html'
             },
             {
                 route: '/categories/income/create',
                 title: 'Создание категории доходов',
                 filePathTemplate: '/templates/pages/category income/create.html',
-                load: () => {
-                    new IncomeCreate();
-                },
+                load: () => new CategoryCreate(this.openNewRoute.bind(this), 'income'),
                 useLayout: '/templates/layout.html'
             },
             {
                 route: '/categories/income/edit',
                 title: 'Редактирование категории доходов',
                 filePathTemplate: '/templates/pages/category income/edit.html',
-                load: () => {
-                    new IncomeEdit(this.openNewRoute.bind(this));
-                },
+                load: () => new CategoryEdit(this.openNewRoute.bind(this), 'income'),
                 useLayout: '/templates/layout.html'
             },
             {
                 route: '/categories/expense',
                 title: 'Категория расходов',
                 filePathTemplate: '/templates/pages/category expense/list.html',
-                load: () => {
-                    new ExpenseList(this.openNewRoute.bind(this));
-                },
+                load: () => new CategoryList(this.openNewRoute.bind(this), 'expense'),
                 useLayout: '/templates/layout.html'
             },
             {
                 route: '/categories/expense/create',
                 title: 'Создание категории расходов',
                 filePathTemplate: '/templates/pages/category expense/create.html',
-                load: () => {
-                    new ExpenseCreate(this.openNewRoute.bind(this));
-                },
+                load: () => new CategoryCreate(this.openNewRoute.bind(this), 'expense'),
                 useLayout: '/templates/layout.html'
             },
             {
                 route: '/categories/expense/edit',
                 title: 'Редактирование категории расходов',
                 filePathTemplate: '/templates/pages/category expense/edit.html',
-                load: () => {
-                    new ExpenseEdit(this.openNewRoute.bind(this));
-                },
+                load: () => new CategoryEdit(this.openNewRoute.bind(this), 'expense'),
                 useLayout: '/templates/layout.html'
             },
             {
                 route: '/operations',
                 title: 'Операции',
                 filePathTemplate: '/templates/pages/operations/list.html',
-                load: () => {
-                    new OperationList(this.openNewRoute.bind(this));
-                },
+                load: () => new OperationList(this.openNewRoute.bind(this)),
                 useLayout: '/templates/layout.html'
             },
             {
                 route: '/operations/create',
                 title: 'Создание операции',
                 filePathTemplate: '/templates/pages/operations/create.html',
-                load: () => {
-                    new OperationCreate();
-                },
+                load: () => new OperationCreate(this.openNewRoute.bind(this)),
                 useLayout: '/templates/layout.html'
             },
             {
                 route: '/operations/edit',
                 title: 'Обновление операции',
                 filePathTemplate: '/templates/pages/operations/edit.html',
-                load: () => {
-                    new OperationEdit(this.openNewRoute.bind(this));
-                },
+                load: () => new OperationEdit(this.openNewRoute.bind(this)),
                 useLayout: '/templates/layout.html'
-            },
-            {
-                route: '/operations/delete',
-                load: () => {
-                    new OperationDelete();
-                }
             },
             {
                 route: '/login',
@@ -154,9 +123,7 @@ export class Router {
             },
             {
                 route: '/logout',
-                load: () => {
-                    new Logout(this.openNewRoute.bind(this));
-                }
+                load: () => new Logout(this.openNewRoute.bind(this))
             },
             {
                 route: '/404',
@@ -176,7 +143,8 @@ export class Router {
     }
 
     async openNewRoute(url) {
-        if (!AuthUtils.getAuthInfo(AuthUtils.accessTokenKey) || !AuthUtils.getAuthInfo(AuthUtils.refreshTokenKey)) {
+        if ((!AuthUtils.getAuthInfo(AuthUtils.accessTokenKey) || !AuthUtils.getAuthInfo(AuthUtils.refreshTokenKey))
+            && url !== '/signup') {
             localStorage.clear();
             url = '/login';
         }
@@ -236,17 +204,7 @@ export class Router {
                 }
                 contentBlock.innerHTML = await fetch(newRoute.filePathTemplate).then(response => response.text());
 
-                let userInfo = AuthUtils.getAuthInfo(AuthUtils.userInfoTokenKey);
-                if (userInfo) {
-                    userInfo = JSON.parse(userInfo);
-                    if (userInfo.name && userInfo.lastName && userInfo.balance.toString()) {
-                        document.getElementById('person-name').innerText =
-                            (!userInfo.name && !userInfo.lastName) ? 'Пользователь' : userInfo.name + ' ' + userInfo.lastName;
-                        document.getElementById('balance-number').innerText =
-                            (userInfo.balance === 0) ? '0$' : userInfo.balance + '$';
-                    }
-                }
-
+                await this.setUserBalance(AuthUtils.getAuthInfo(AuthUtils.accessTokenKey), AuthUtils.getAuthInfo(AuthUtils.refreshTokenKey));
             }
             if (newRoute.load && typeof newRoute.load === "function") {
                 newRoute.load();
@@ -257,5 +215,25 @@ export class Router {
             history.pushState({}, '', '/404');
             await this.activateRoute();
         }
+    }
+
+    async setUserBalance(accessToken, refreshToken) {
+        const result = await HttpUtils.request('/balance');
+
+        if (result.error || !result.response) {
+            console.log(result.response.message)
+            return false;
+        }
+
+        const userInfo = JSON.parse(AuthUtils.getAuthInfo(AuthUtils.userInfoTokenKey));
+        if (userInfo) {
+            userInfo.balance = result.response.balance;
+            AuthUtils.setAuthInfo(null, null, JSON.stringify(userInfo));
+        }
+
+        document.getElementById('person-name').innerText =
+            (!userInfo.name && !userInfo.lastName) ? 'Пользователь' : userInfo.name + ' ' + userInfo.lastName;
+        document.getElementById('balance-number').innerText =
+            (userInfo.balance === 0) ? '0$' : userInfo.balance + '$';
     }
 }
